Endpoint /videos using media-api

+++ src/resources/videos.py
@@ -1,13 +1,14 @@
 import datetime
 from http import HTTPStatus
-from flask import make_response, request, g
+from flask import make_response, request, g, current_app as app
 from flask_restful import Resource
 from marshmallow import ValidationError
+from src.clients.media_api import MediaAPIClient
 from src.misc.authorization import check_token
 from src.misc.responses import response_error, response_ok
 from src.models.comment import Comment
 from src.models.video import Video
-from src.schemas.video import VideoSchema, VideoPaginatedSchema
+from src.schemas.video import VideoSchema, VideoPaginatedSchema, MediaSchema
 
 
 class Videos(Resource):
@@ -15,7 +16,14 @@ class Videos(Resource):
     def get(self, video_id):
         schema = VideoSchema()
         video = Video.objects(id=video_id).first_or_404()
-        return make_response(schema.dump(video), HTTPStatus.OK)
+        resp_media = MediaAPIClient.get_video(video.id)
+        if resp_media.status_code != HTTPStatus.OK:
+            app.logger.error("[video_id:%s] Error getting media from media-server: %s" %
+                             (video.id, resp_media.json()["message"]))
+            return response_error(HTTPStatus.INTERNAL_SERVER_ERROR, "Error getting media")
+        result = schema.dump(video)
+        result["media"] = resp_media.json()
+        return make_response(result, HTTPStatus.OK)
 
     @check_token
     def put(self, video_id):
@@ -33,13 +41,25 @@ class Videos(Resource):
         video.visibility = new_video.visibility
         video.date_updated = datetime.datetime.utcnow()
         video.save()
-        return make_response(schema.dump(video), HTTPStatus.OK)
+        resp_media = MediaAPIClient.get_video(video.id)
+        if resp_media.status_code != HTTPStatus.OK:
+            app.logger.error("[video_id:%s] Error getting media from media-server: %s" %
+                             (video.id, resp_media.json()["message"]))
+            return response_error(HTTPStatus.INTERNAL_SERVER_ERROR, "Error getting media")
+        result = schema.dump(video)
+        result["media"] = resp_media.json()
+        return make_response(result, HTTPStatus.OK)
 
     @check_token
     def delete(self, video_id):
         video = Video.objects(id=video_id).first_or_404()
         if video.user != g.session_username:
             return response_error(HTTPStatus.FORBIDDEN, str("Forbidden"))
+        resp_media = MediaAPIClient.delete_video(video.id)
+        if resp_media.status_code != HTTPStatus.OK:
+            app.logger.error("[video_id:%s] Error deleting media from media-server: %s" %
+                             (video.id, resp_media.json()["message"]))
+            return response_error(HTTPStatus.INTERNAL_SERVER_ERROR, "Error getting media")
         video.delete()
         Comment.objects(video=video_id).delete()
         return response_ok(HTTPStatus.OK, "Video deleted")
@@ -50,19 +70,41 @@ class VideosList(Resource):
     def get(self):
         schema = VideoPaginatedSchema()
         paginated = schema.load(request.args)
-        video = Video.objects(**paginated["filters"]).skip(paginated["offset"]).limit(paginated["limit"])
-        return make_response(dict(data=VideoSchema().dump(video, many=True)), HTTPStatus.OK)
+        videos = Video.objects(**paginated["filters"]).skip(paginated["offset"]).limit(paginated["limit"])
+        results = []
+        for video in videos:
+            resp_media = MediaAPIClient.get_video(video.id)
+            if resp_media.status_code != HTTPStatus.OK:
+                app.logger.error("[video_id:%s] Error getting media from media-server: %s" %
+                                 (video.id, resp_media.json()["message"]))
+                continue
+            result = schema.dump(video)
+            result["media"] = resp_media.json()
+            results.append(result)
+        return make_response(dict(data=results), HTTPStatus.OK)
 
     @check_token
     def post(self):
-        schema = VideoSchema()
+        schema_video = VideoSchema()
+        schema_media = MediaSchema()
         try:
-            video = schema.load(request.get_json(force=True))
+            request_json = request.get_json(force=True)
+            video = schema_video.load(request_json)
+            media = schema_media.load(request_json.get("media", {}))
         except ValidationError as e:
             return response_error(HTTPStatus.BAD_REQUEST, str(e.normalized_messages()))
         now = datetime.datetime.utcnow()
         video.user = g.session_username
         video.date_created = now
         video.date_updated = now
-        video.save()
-        return make_response(schema.dump(video), HTTPStatus.CREATED)
+        new_video = video.save()
+        media["video_id"] = new_video.id
+        res = MediaAPIClient.post_video(schema_media.dump(media))
+        if res.status_code != HTTPStatus.OK:
+            app.logger.error("[video_name:%s] Error saving new video on media-server: %s" %
+                             (media["name"], res.json()["message"]))
+            video.delete()
+            return response_error(HTTPStatus.INTERNAL_SERVER_ERROR, "Error saving media")
+        result = schema_video.dump(new_video)
+        result["media"] = res.json()
+        return make_response(result, HTTPStatus.CREATED)
